refactor: app 설정
