+++ src/app/App.tsx
@@ -1,0 +1,12 @@
+import PostManagerPage from "@pages/PostsManagerPage"
+import { BrowserRouter as Router } from "react-router-dom"
+
+const App = () => {
+  return (
+    <Router>
+      <PostManagerPage />
+    </Router>
+  )
+}
+
+export default App
