+++ src/widgets/Layout.tsx
@@ -1,0 +1,14 @@
+import Header from "@widgets/ui/Header.tsx"
+import Footer from "@widgets/ui/Footer.tsx"
+
+const Layout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div className="flex flex-col min-h-screen">
+      <Header />
+      <main className="flex-grow container mx-auto px-4 py-8">{children}</main>
+      <Footer />
+    </div>
+  )
+}
+
+export default Layout
